Add files via upload
The data is provided by a cab aggregator service called SIGMA CABS and the surge type price has been captured. The main business objective is to build a predictive model, which could help in predicting the surge pricing type pro-actively. This, in turn, would help them in matching the right cabs with the right customers quickly and efficiently.